DeepResearchAgent/pyproject.toml
[project]
name = "deepresearchagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "openai-agents>=0.2.10",
    "tavily-python>=0.7.11",
]
